+++ docs/astro.config.mjs
@@ -58,8 +58,9 @@ export default defineConfig({
         },
       ],
       editLink: {
-        baseUrl: "https://github.com/brycefrank/pyfor/edit/master/docs/",
+        baseUrl: "https://github.com/brycefrank/pyfor/edit/main/docs/",
       },
+      customCss: ["./src/styles/custom.css"],
       sidebar: [
         { label: "Introduction", link: "/introduction/" },
         { label: "Installation", link: "/installation/" },